STOPGAP: Load Cloudflare origin certificates in code under Production environment
Because this app uses Nancy, it doesn't follow the standard conventions... so if we add the Kestrel config to the config.production.json it just crashes out (we're only expecting single values).

So, we have to define the Kestrel options in code here, and get the certificate details from the environment.

Locally (in the Development environment), this just runs as normal, using the ASP.NET dev certificate.